Fixed compatibility with 0.8.6
The clampNumber method seems to be gone in this version of foundry. This has the effect of not allowing a specific number of items to be traded, just the max. Changed trade-window.js to use a couple if statements instead.

trade-request.js just added some additional verification in the isValid method

trade-window.html I just changed the default quantity to trade to 1 instead of the max.